+++ frontend/customerapp/lib/screens/products/products.dart
@@ -9,6 +9,8 @@ import 'package:customerapp/models/products.dart';
 import 'package:customerapp/models/restaurants.dart';
 import 'package:customerapp/screens/commonComponents/single_message_dialog.dart';
 import 'package:customerapp/screens/products/cart_box.dart';
+import 'package:customerapp/styles/Komet.dart';
+import 'package:customerapp/styles/category_prod.dart';
 import 'package:customerapp/styles/product.dart';
 import 'package:customerapp/responsive/screen_responsive.dart';
 import 'package:flutter/material.dart';
@@ -123,12 +125,21 @@ class _Products extends State<Products> {
                                                       color: Color(0xAAFFFFFF)),
                                                   alignment:
                                                       Alignment.centerLeft,
-                                                  child: Text(
-                                                    restaurant == null
-                                                        ? 'Products'
-                                                        : restaurant.name,
-                                                    style: restaurantTitleStyle,
-                                                  ),
+                                                  child: Column(children: [
+                                                    Padding(
+                                                        padding:
+                                                            EdgeInsets.only(
+                                                                top: 10),
+                                                        child: Text(
+                                                          restaurant == null
+                                                              ? 'Products'
+                                                              : restaurant.name,
+                                                          style:
+                                                              restaurantTitleStyle,
+                                                        )),
+                                                    Divider(),
+                                                    CustomRadio()
+                                                  ]),
                                                 ),
                                               ),
                                               Expanded(
@@ -248,4 +259,145 @@ class _Products extends State<Products> {
           }
         },
       );
+}
+
+class CustomRadio extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() {
+    return CustomRadioState();
+  }
+}
+
+class CustomRadioState extends State<CustomRadio> {
+  List<RadioModel> sample = List<RadioModel>();
+
+  List<Widget> generate_widgets() {
+    List<Widget> llista = List<Widget>();
+    sample.asMap().forEach((key, value) {
+      llista.add(InkWell(
+          onTap: () {
+            setState(() {
+              sample.forEach((element) => element.isSelected = false);
+              sample[key].isSelected = true;
+            });
+          },
+          child:
+              Product_class_widget(sample[key].name, sample[key].isSelected)));
+    });
+
+    return llista;
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    sample.add(RadioModel("primer", true));
+    sample.add(RadioModel("segundo", false));
+    sample.add(RadioModel("tercer", false));
+    sample.add(RadioModel("quart", false));
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: generate_widgets(),
+    );
+  }
+}
+
+class Product_class_widget extends StatelessWidget {
+  String name;
+  bool isSelected;
+
+  Product_class_widget(this.name, this.isSelected);
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+        padding: EdgeInsets.only(left: 10),
+        child: Container(
+            child: Column(mainAxisSize: MainAxisSize.min, children: [
+          Container(
+              child: Text(name,
+                  style: isSelected
+                      ? ProdTextTitleStyle
+                      : ProdTextTitleStyle_basic)),
+          Container(
+            height: isSelected ? 7 : 0,
+            width: name.length.toDouble() * 10,
+            decoration: BoxDecoration(
+                color: kommetDistinctiveYellow,
+                borderRadius: BorderRadius.all(Radius.circular(40))),
+          )
+        ])));
+  }
+}
+
+class RadioModel {
+  String name;
+  bool isSelected;
+
+  RadioModel(this.name, this.isSelected);
+}
+
+class GridView_general extends StatelessWidget {
+  var restaurant, productsModel;
+
+  Cart cart;
+  bool firstInstance = true;
+
+  GridView_general(this.restaurant, this.productsModel);
+
+  void addToCart(Product prod) {
+    cart.addItem(prod);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    cart = context.watch<Cart>();
+    if (firstInstance) {
+      productsModel.removeProducts();
+      cart.empty();
+    }
+
+    return Expanded(
+      child: Builder(builder: (builder) {
+        return FutureBuilder(
+          future: getProductsFromRestaurant(restaurant.id),
+          builder: (BuildContext context, AsyncSnapshot snapshot) {
+            if (snapshot.hasData) {
+              productsModel.removeProducts();
+              snapshot.data.forEach((element) {
+                productsModel.addProduct(Product.fromDTO(element));
+              });
+              return StaggeredGridView.countBuilder(
+                crossAxisSpacing: 10,
+                mainAxisSpacing: 10,
+                physics:
+                    NeverScrollableScrollPhysics(), // to disable GridView's scrolling
+                shrinkWrap: true,
+                itemCount: productsModel.availableProducts.length,
+                crossAxisCount: MediaQuery.of(context).size.width > 600 ? 2 : 1,
+                itemBuilder: (context, index) {
+                  return ProductListCard(Key('product-card-$index'),
+                      productsModel.availableProducts[index], addToCart);
+                },
+                staggeredTileBuilder: (int index) => StaggeredTile.fit(1),
+              );
+            } else {
+              return Padding(
+                  padding: EdgeInsets.fromLTRB(0, 0, 0, 10),
+                  child: Container(
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(15.0),
+                        color: Color(0xAAFFFFFF)),
+                    alignment: Alignment.centerLeft,
+                    child: CircularLoaderKomet(),
+                  ));
+            }
+          },
+        );
+      }),
+    );
+  }
 }
